Large changes to theming and file layout.
- Theming absracted and to allow a standard theme of dark/light
and all interfaces broken down ready to be extracted out of styled
components to create additional theming for users.
- file structured amended to use a standard way of structure: using
an index.tsx file for the base React component, a props.ts holding
the interface for said component, a styled.ts file for styled components
and a stories directory for all stories.
zsh:1: command not found: wq